requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==9.5.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
django-redis==5.2.0
drf-yasg==1.21.7
python-decouple==3.8
django-cleanup==7.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7